Mudando estrutura e polindo
