+++ assets/js/Calculadora.js
@@ -1,4 +1,4 @@
-class Calculadora {
+export default class Calculadora {
     sumar(num1, num2) {
         return num1 + num2;
     }
