+++ index.html
@@ -1,14 +1,16 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="pt-br">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Calculadora</title>
+
+    <!-- Google fonts -->
     <link href="https://fonts.googleapis.com/css2?family=Fraunces&display=swap" rel="stylesheet">
-    <link href="index.css" rel="stylesheet" />
-    <script src="Display.js" type="text/javascript"></script>
-    <script src="Calculadora.js" type="text/javascript"></script>
-    <script src="index.js" type="text/javascript" defer></script>
+    <!-- css -->
+    <link href="./assets/css/styles.css" rel="stylesheet" />
+    <!-- js -->
+    <script src="./assets/js/index.js" type="module" defer></script>
 </head>
 <body>
     <div class="container">
@@ -17,8 +19,8 @@
                 <div id="valor-anterior"></div>
                 <div id="valor-actual"></div>
             </div>
-            <button class="col-2" onclick="display.borrarTodo()">C</button>
-            <button onclick="display.borrar()">&larr;</button>
+            <button class="col-2" data-func="borrar-todo">C</button>
+            <button data-func="borrar">&larr;</button>
             <button class="operador" value="dividir">%</button>
             <button class="numero">7</button>
             <button class="numero">8</button>
